expose RenderInlineModel
