+++ src/frontend/src/components/render/Build.tsx
@@ -1,8 +1,9 @@
 import type { ReactNode } from 'react';
 
+import { RenderInlineModel } from '@lib/components/RenderInlineModel';
 import { ModelType } from '@lib/enums/ModelType';
 import { getDetailUrl } from '@lib/functions/Navigation';
-import { type InstanceRenderInterface, RenderInlineModel } from './Instance';
+import type { InstanceRenderInterface } from './Instance';
 import { StatusRenderer } from './StatusRenderer';
 
 /**
