+++ src/frontend/src/components/render/Company.tsx
@@ -2,9 +2,10 @@ import type { ReactNode } from 'react';
 
 import { Text } from '@mantine/core';
 
+import { RenderInlineModel } from '@lib/components/RenderInlineModel';
 import { ModelType } from '@lib/enums/ModelType';
 import { getDetailUrl } from '@lib/functions/Navigation';
-import { type InstanceRenderInterface, RenderInlineModel } from './Instance';
+import type { InstanceRenderInterface } from './Instance';
 
 /**
  * Inline rendering of a single Address instance
