+++ src/frontend/src/components/render/Part.tsx
@@ -2,13 +2,14 @@ import { t } from '@lingui/core/macro';
 import { Badge, Group, Text } from '@mantine/core';
 import type { ReactNode } from 'react';
 
+import { RenderInlineModel } from '@lib/components/RenderInlineModel';
 import { ModelType } from '@lib/enums/ModelType';
 import { formatDecimal } from '@lib/functions/Formatting';
 import { getDetailUrl } from '@lib/functions/Navigation';
 import { shortenString } from '@lib/functions/String';
 import { TableHoverCard } from '../../tables/TableHoverCard';
 import { ApiIcon } from '../items/ApiIcon';
-import { type InstanceRenderInterface, RenderInlineModel } from './Instance';
+import type { InstanceRenderInterface } from './Instance';
 
 /**
  * Inline rendering of a single Part instance
